fix path creation bug and evaluation of arguments
If path did not exist, and independent of the "-m" flag, script won't work
because the use of realpath on the declaration of $path just fails if a
non-existent directory is given. This is easily solvable by just creating the
path (if "-m" flag present) previous to the $path declaration.

Also, some crazy annoyances were given while evaluating the shorthand
conditional expressions on the existence of flags. Seems that the expr:

    [[ condition ]] && do_something

returns a non-zero value when condition is false, unlike a normal `if`. So they
were all replaced by if conditionals.